Create classy-multi-az-ubuntu-network-calico.yaml
ClusterClass manifest for multi-AZ deployments across 3 Zones. 

Uses Ubuntu images which must be present in the Content Library associated with the vSphere Namespace where this cluster deploys (e.g. cormac-ns)

Includes a ClusterBootstrap to use Calico CNI instead of Antrea CNI. This requires the KappController which is also included

Note that cidrBlocks *must* be specified in ClusterClass YAMLs - there is no webhook to cater for default settings.