Create classy-az-single-fd-with-pvs.yaml
Deploys a ClusterClass into a multi-AZ setup, but enforces that a fault domain is chosen for the worker node.

The example uses variables to persistent volumes to control plane and worker nodes i.e., etcd, containerd, and kubelet. vmClass variable is declared twice. The vmClass variable declared in workers.machineDeployments overwrites the globally declared vmClass variable so that worker nodes are provisioned with a different VM class if necessary.